generic update tracker
